# Annotate some cold functions as cold, noinline